Fix migrations
- Merge some migrations
- Remove empty migrations
- Add missing field collection index fields
- Format migrations
