--- pb_migrations/1750750200_add_zone_to_page_type_sections.js
@@ -1,11 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate(
-	(app) => {
-		// Empty migration - zone field will be added manually through admin UI
-		return null
-	},
-	(app) => {
-		// Empty rollback
-		return null
-	}
-)
