+++ pb_migrations/1750751803_updated_page_type_sections.js
@@ -1,30 +1,48 @@
 /// <reference path="../pb_data/types.d.ts" />
-migrate((app) => {
-  const collection = app.findCollectionByNameOrId("pbc_496957621")
+migrate(
+	(app) => {
+		const collection = app.findCollectionByNameOrId('pbc_496957621')
 
-  // add field
-  collection.fields.addAt(4, new Field({
-    "hidden": false,
-    "id": "select2699804679",
-    "maxSelect": 1,
-    "name": "zone",
-    "presentable": false,
-    "required": false,
-    "system": false,
-    "type": "select",
-    "values": [
-      "header",
-      "content",
-      "footer"
-    ]
-  }))
+		// update collection data
+		unmarshal(
+			{
+				indexes: ['CREATE UNIQUE INDEX `idx_18mJGVnXFu` ON `page_type_sections` (\n  `page_type`,\n  `index`,\n  `zone`\n)']
+			},
+			collection
+		)
 
-  return app.save(collection)
-}, (app) => {
-  const collection = app.findCollectionByNameOrId("pbc_496957621")
+		// add field
+		collection.fields.addAt(
+			4,
+			new Field({
+				hidden: false,
+				id: 'select2699804679',
+				maxSelect: 1,
+				name: 'zone',
+				presentable: false,
+				required: true,
+				system: false,
+				type: 'select',
+				values: ['header', 'body', 'footer']
+			})
+		)
 
-  // remove field
-  collection.fields.removeById("select2699804679")
+		return app.save(collection)
+	},
+	(app) => {
+		const collection = app.findCollectionByNameOrId('pbc_496957621')
 
-  return app.save(collection)
-})
+		// update collection data
+		unmarshal(
+			{
+				indexes: ['CREATE UNIQUE INDEX `idx_18mJGVnXFu` ON `page_type_sections` (\n  `page_type`,\n  `index`\n)']
+			},
+			collection
+		)
+
+		// remove field
+		collection.fields.removeById('select2699804679')
+
+		return app.save(collection)
+	}
+)
